A lot more fixes.

+++ src/main/java/plus/dragons/createenchantmentindustry/foundation/mixin/FluidTankBlockMixin.java
@@ -1,5 +1,10 @@
 package plus.dragons.createenchantmentindustry.foundation.mixin;
 
+import org.spongepowered.asm.mixin.Mixin;
+import org.spongepowered.asm.mixin.injection.At;
+import org.spongepowered.asm.mixin.injection.Inject;
+import org.spongepowered.asm.mixin.injection.callback.CallbackInfo;
+
 import com.simibubi.create.api.connectivity.ConnectivityHandler;
 import com.simibubi.create.content.equipment.wrench.IWrenchable;
 import com.simibubi.create.content.fluids.tank.CreativeFluidTankBlockEntity;
@@ -8,16 +13,13 @@ import com.simibubi.create.content.fluids.tank.FluidTankBlockEntity;
 import com.simibubi.create.content.processing.basin.BasinBlockEntity;
 import com.simibubi.create.foundation.block.IBE;
 import com.simibubi.create.foundation.utility.VecHelper;
+
 import net.minecraft.core.BlockPos;
 import net.minecraft.server.level.ServerLevel;
 import net.minecraft.world.level.Level;
 import net.minecraft.world.level.block.Block;
 import net.minecraft.world.level.block.entity.BlockEntity;
 import net.minecraft.world.level.block.state.BlockState;
-import org.spongepowered.asm.mixin.Mixin;
-import org.spongepowered.asm.mixin.injection.At;
-import org.spongepowered.asm.mixin.injection.Inject;
-import org.spongepowered.asm.mixin.injection.callback.CallbackInfo;
 import plus.dragons.createenchantmentindustry.content.contraptions.fluids.experience.ExperienceFluid;
 
 @Mixin(FluidTankBlock.class)
@@ -42,12 +44,12 @@ public abstract class FluidTankBlockMixin extends Block implements IBE<BasinBloc
             level.removeBlockEntity(pos);
             ConnectivityHandler.splitMulti(tankBE);
             if (maxSize == 1) {
-                expFluid.drop(serverLevel, VecHelper.getCenterOf(pos), fluidStackBackup.getAmount());
+                expFluid.drop(serverLevel, VecHelper.getCenterOf(pos), (int) fluidStackBackup.getAmount());
             } else {
                 var total = maxSize * (FluidTankBlockEntity.getCapacityMultiplier() - 1);
                 var leftover = fluidStackBackup.getAmount() - total;
                 if(leftover > 0) {
-                    expFluid.drop(serverLevel, VecHelper.getCenterOf(pos), leftover);
+                    expFluid.drop(serverLevel, VecHelper.getCenterOf(pos), (int) leftover);
                 }
             }
             ci.cancel();
